font: use regular JB Mono with dedicated Symbol font
Fixes #7069

I'm not sure if this will have follow-on negative effects. In my day to
day usage everything looks fine but its hard to really know. I think our
best chance is to merge this and see what feedback we get. I know for
sure positively that it does two things:

  1. Fixes the small symbols noted in #7069.
  2. Makes our binary size smaller since we no longer have unnecessary
     symbols embedded in our stylized fonts.